fix(#2506): convert aerodynamics.py to package to reduce LOC per module
aerodynamics.py was 1135 LOC (budget: 1200, target: 300). Converted to a
package with five focused modules — _config, _models, _wind, _environment,
_engine — each under 300 LOC. Public API re-exported from __init__.py for
backward compatibility; all 93 existing aerodynamics tests pass unchanged.
